+++ src-docs/src/views/header/header_alert.js
@@ -277,7 +277,6 @@ const HeaderUserMenu = () => {
   return (
     <EuiPopover
       id={id}
-      ownFocus
       repositionOnScroll
       button={button}
       isOpen={isOpen}