+++ src/components/date_picker/super_date_picker/quick_select_popover/quick_select_popover.tsx
@@ -186,8 +186,7 @@ export class EuiQuickSelectPopover extends Component<
         isOpen={isOpen}
         closePopover={this.closePopover}
         anchorPosition="downLeft"
-        anchorClassName="euiQuickSelectPopover__anchor"
-        ownFocus>
+        anchorClassName="euiQuickSelectPopover__anchor">
         <div
           className="euiQuickSelectPopover__content"
           data-test-subj="superDatePickerQuickMenu">