+++ src/components/header/header_links/header_links.tsx
@@ -138,7 +138,6 @@ export const EuiHeaderLinks: FunctionComponent<EuiHeaderLinksProps> = ({
 
           <EuiShowFor sizes={popoverBreakpoints}>
             <EuiPopover
-              ownFocus
               button={button}
               isOpen={mobileMenuIsOpen}
               anchorPosition="downRight"